Constructor for fi working(creating hs and rs), fixed distribution functions, in lsh now dimensions and lines of file get computed before creating fi

+++ help_functions.h
@@ -3,6 +3,6 @@
 
 namespace help_functions
 {
-	float * normal_distribution_vector(int dim);
+	double * normal_distribution_vector(int dim);
 }
 #endif
